backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
passlib[bcrypt]
PyJWT
pytest
httpx
ruff
